--- src/main/java/hanium/dtc/mypage/repository/MyPageRepository.java
@@ -1,7 +1,0 @@
-package hanium.dtc.mypage.repository;
-
-import hanium.dtc.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MyPageRepository extends JpaRepository<User, Long> {
-}
